Write loop that creates div based on getTiles()
The function createCanvas() takes the array created in getTiles()
and creates a div for each of the items inside the div #canvas.
Having both steps (calculating number of tiles needed and creating
the appropriate number of div) being done through loops should enable
me to adjust the canvas size easily on a later point.