Implement buttons to determine 'paint color'
I create three buttons to give the user the option to use a paint color
(black, red and white for "erasing").
For the buttons to have effect I replaced the color setting in the
code responsible for drawing through a variable, which is set through
eventListener on the color buttons.