Create function to set tile size fro grit
Instead of manually setting the width and height of the tiles in the
grit, I created a function that calculates the width/height based on the
grit size. That makes it possible to now implement a way for the user
to set their own grit size and the tile size will adjust automatically.