Introduce option to set grit size through buttons
I implemented buttons that let user choose from a preset of grit sizes.
If pressed the program deletes the initial tiles and fills canvas with
the adjusted amount of new tiles, therefor making the canvas resolution
bigger or smaller.
For this to work correctly I had to make various changes on the
existing code, mainly create function for some of the canvas creation
steps and including the variable declaration for 'tiles', so that I can
repeatably access them and the value is updated with each new grit
creation.