Change mode of setting initial backgroundColor
I changed the initial 'coloring' of the canvas so that the application
starts with giving each element a backgroundColor value through java
instead if css styling. This way there is a rgb value that can be
accessed when using the darkening feature on 'empty' tile.
I also regrouped some of the code for readability.