Automate pulling canvas measurements for tiles
Instead of having to set the width and height of the canvas manually
to calculate the appropriate size for the tiles I wanted the function
to retrieve the pixel size directly from the css and adjust the tile
sizes automatically if the size in the css file changes.
It will make it easier to implement features like making the canvas
bigger on the screen or make the site responsive.

Further I fixed a bug, were the grit would vanish when the grit size
is changed.